feat: refactor App.jsx following React best practices
Major improvements:
- ✅ Component breakdown: Split 623-line monolithic App into 10+ focused components
- ✅ Custom hooks: Extract state management (useTheme, useFileUpload, useReviewData, usePagination)
- ✅ Business logic separation: Move utilities to dedicated files (mockDataGenerator, tooltipPositioning)
- ✅ Performance optimization: Add React.memo, useCallback for all components
- ✅ Error boundaries: Add ErrorBoundary component for graceful error handling
- ✅ Type safety: Add PropTypes for runtime validation
- ✅ Accessibility: Maintain and enhance existing a11y features
- 🐛 Fix: Restore review section fun message display before submit

New file structure:
- components/: Header, FileUpload, ReviewSection, LoadingState, SearchBar, etc.
- hooks/: useTheme, useFileUpload, useReviewData, usePagination
- utils/: mockDataGenerator, tooltipPositioning

App.jsx reduced from 623 to 216 lines with improved maintainability and performance.
