+++ apps/ratio-ui/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "papaparse": "^5.5.3",
+    "prop-types": "^15.8.1",
     "react": "^19.1.1",
     "react-dom": "^19.1.1"
   },
