+++ apps/ratio-ui/src/App.css
@@ -127,8 +127,8 @@
     --focus-offset: 0.125rem; /* 2px */
 
     /* Artist Mode Cursors */
-    --artist-brush-cursor: url("data:image/svg+xml;charset=utf8,%3Csvg xmlns='http://www.w3.org/2000/svg' width='69' height='103' viewBox='0 0 69 103' fill='none'%3E%3Cpath d='M16.3233 7.57656L9.06567 12.0874C7.51167 8.19816 5.75607 5.01456 3.32007 1.76376C8.26767 1.20096 13.4337 3.53616 16.3233 7.57656Z' fill='white' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3Cpath d='M66.0932 94.3648C64.262 97.3552 60.5156 97.1536 56.8952 97.1788C48.6212 97.246 41.9432 91.618 41.9432 91.618C41.9432 91.618 42.8336 96.2296 43.85 98.4052C38.6084 94.9024 33.686 90.9292 29.1416 86.5612C29.5028 89.1064 29.864 91.6432 30.2252 94.1884C23.0012 87.3676 15.6176 80.2696 10.04 71.9872C7.7048 68.518 5.924 65.3848 5.462 61.21C4.8824 56.0272 4.7648 50.7856 5.1176 45.586C5.4788 40.252 6.9908 34.0444 13.7192 34.0108C13.9544 34.0108 14.1644 34.0108 14.366 34.0108L34.778 33.1372C37.676 33.448 40.4732 34.5652 42.7832 36.3376C47.5376 35.5816 52.586 37.4716 56.1308 40.7392C59.6672 43.9984 61.9268 48.4672 63.2792 53.0872C64.6232 57.7072 65.1272 62.5372 65.6312 67.3252C66.0932 71.8024 66.5552 76.2964 67.0172 80.7736C67.4204 84.6376 68.126 91.0552 66.0932 94.3648Z' fill='white' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3Cpath d='M14.3662 34.0031C17.113 33.9779 17.6338 33.9359 20.7166 34.3727' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3Cpath d='M42.7832 36.3296C43.2032 36.6488 43.6148 36.9848 43.9928 37.3544C46.6136 39.8156 48.3608 43.184 48.8648 46.7372' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3Cpath d='M34.7865 33.129C34.6353 33.1038 34.4841 33.0954 34.3413 33.0786' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3Cpath d='M20.4392 63.7051C20.2628 59.6479 20.456 58.6231 20.54 55.9855L16.4576 57.6739C17.5832 53.8267 17.936 46.5691 17.1632 42.6043C16.466 38.9755 15.7436 37.4635 14.3744 34.0111C14.0132 33.0787 13.6268 32.1463 13.2236 31.2307C11.6192 27.5599 10.1828 23.8051 9.12441 19.9411C8.21721 16.6483 7.69641 11.4151 10.3172 8.59265C10.8296 8.03825 11.468 7.57625 12.2492 7.24025C13.61 6.66065 14.9036 6.49265 16.1132 6.63545C18.7088 6.93785 20.96 8.64305 22.766 10.7767C28.352 17.3455 32.4008 24.9643 34.7864 33.1375C35.9372 37.0687 37.3988 40.4119 37.76 44.5447' fill='white'/%3E%3Cpath d='M20.4392 63.7051C20.2628 59.6479 20.456 58.6231 20.54 55.9855L16.4576 57.6739C17.5832 53.8267 17.936 46.5691 17.1632 42.6043C16.466 38.9755 15.7436 37.4635 14.3744 34.0111C14.0132 33.0787 13.6268 32.1463 13.2236 31.2307C11.6192 27.5599 10.1828 23.8051 9.12441 19.9411C8.21721 16.6483 7.69641 11.4151 10.3172 8.59265C10.8296 8.03825 11.468 7.57625 12.2492 7.24025C13.61 6.66065 14.9036 6.49265 16.1132 6.63545C18.7088 6.93785 20.96 8.64305 22.766 10.7767C28.352 17.3455 32.4008 24.9643 34.7864 33.1375C35.9372 37.0687 37.3988 40.4119 37.76 44.5447' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3C/svg%3E")
-      20 9;
+    --artist-brush-cursor: url("data:image/svg+xml;charset=utf8,%3Csvg xmlns='http://www.w3.org/2000/svg' width='35' height='52' viewBox='0 0 69 103' fill='none'%3E%3Cpath d='M16.3233 7.57656L9.06567 12.0874C7.51167 8.19816 5.75607 5.01456 3.32007 1.76376C8.26767 1.20096 13.4337 3.53616 16.3233 7.57656Z' fill='white' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3Cpath d='M66.0932 94.3648C64.262 97.3552 60.5156 97.1536 56.8952 97.1788C48.6212 97.246 41.9432 91.618 41.9432 91.618C41.9432 91.618 42.8336 96.2296 43.85 98.4052C38.6084 94.9024 33.686 90.9292 29.1416 86.5612C29.5028 89.1064 29.864 91.6432 30.2252 94.1884C23.0012 87.3676 15.6176 80.2696 10.04 71.9872C7.7048 68.518 5.924 65.3848 5.462 61.21C4.8824 56.0272 4.7648 50.7856 5.1176 45.586C5.4788 40.252 6.9908 34.0444 13.7192 34.0108C13.9544 34.0108 14.1644 34.0108 14.366 34.0108L34.778 33.1372C37.676 33.448 40.4732 34.5652 42.7832 36.3376C47.5376 35.5816 52.586 37.4716 56.1308 40.7392C59.6672 43.9984 61.9268 48.4672 63.2792 53.0872C64.6232 57.7072 65.1272 62.5372 65.6312 67.3252C66.0932 71.8024 66.5552 76.2964 67.0172 80.7736C67.4204 84.6376 68.126 91.0552 66.0932 94.3648Z' fill='white' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3Cpath d='M14.3662 34.0031C17.113 33.9779 17.6338 33.9359 20.7166 34.3727' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3Cpath d='M42.7832 36.3296C43.2032 36.6488 43.6148 36.9848 43.9928 37.3544C46.6136 39.8156 48.3608 43.184 48.8648 46.7372' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3Cpath d='M34.7865 33.129C34.6353 33.1038 34.4841 33.0954 34.3413 33.0786' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3Cpath d='M20.4392 63.7051C20.2628 59.6479 20.456 58.6231 20.54 55.9855L16.4576 57.6739C17.5832 53.8267 17.936 46.5691 17.1632 42.6043C16.466 38.9755 15.7436 37.4635 14.3744 34.0111C14.0132 33.0787 13.6268 32.1463 13.2236 31.2307C11.6192 27.5599 10.1828 23.8051 9.12441 19.9411C8.21721 16.6483 7.69641 11.4151 10.3172 8.59265C10.8296 8.03825 11.468 7.57625 12.2492 7.24025C13.61 6.66065 14.9036 6.49265 16.1132 6.63545C18.7088 6.93785 20.96 8.64305 22.766 10.7767C28.352 17.3455 32.4008 24.9643 34.7864 33.1375C35.9372 37.0687 37.3988 40.4119 37.76 44.5447' fill='white'/%3E%3Cpath d='M20.4392 63.7051C20.2628 59.6479 20.456 58.6231 20.54 55.9855L16.4576 57.6739C17.5832 53.8267 17.936 46.5691 17.1632 42.6043C16.466 38.9755 15.7436 37.4635 14.3744 34.0111C14.0132 33.0787 13.6268 32.1463 13.2236 31.2307C11.6192 27.5599 10.1828 23.8051 9.12441 19.9411C8.21721 16.6483 7.69641 11.4151 10.3172 8.59265C10.8296 8.03825 11.468 7.57625 12.2492 7.24025C13.61 6.66065 14.9036 6.49265 16.1132 6.63545C18.7088 6.93785 20.96 8.64305 22.766 10.7767C28.352 17.3455 32.4008 24.9643 34.7864 33.1375C35.9372 37.0687 37.3988 40.4119 37.76 44.5447' stroke='black' stroke-width='3' stroke-miterlimit='10' stroke-linecap='round'/%3E%3C/svg%3E")
+      10 5;
 
     /* Common values */
     --font-family-system: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
@@ -839,6 +839,9 @@
     overflow-y: auto;
     scrollbar-width: thin;
     scrollbar-color: var(--primary-green) var(--soft-gray);
+    /* Add padding to bottom to prevent tooltip cutoff */
+    padding-bottom: 4rem;
+    margin-bottom: -4rem;
   }
 
   .table-wrapper.scrollable::-webkit-scrollbar {
@@ -1540,9 +1543,9 @@
     /* Dynamic width based on viewport and content */
     min-width: 12rem; /* 192px - minimum width */
     max-width: clamp(20rem, 40vw, 32rem); /* Responsive: 320px to 40% viewport width, max 512px */
+    z-index: 9999; /* Ensure tooltip appears above all content, including scrollable containers */
     width: max-content; /* Expand to content width within constraints */
     box-shadow: 0 0.25rem 0.75rem rgba(0, 0, 0, 0.25);
-    z-index: 9999; /* Highest priority - always on top */
     opacity: 0;
     visibility: hidden;
     transition: all 0.2s ease;
