+++ apps/ratio-ui/src/components/ErrorBoundary.jsx
@@ -1,0 +1,64 @@
+import React from 'react';
+
+class ErrorBoundary extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { hasError: false, error: null, errorInfo: null };
+  }
+
+  static getDerivedStateFromError(error) {
+    // Update state so the next render will show the fallback UI
+    return { hasError: true };
+  }
+
+  componentDidCatch(error, errorInfo) {
+    // Log the error details for debugging
+    console.error('ErrorBoundary caught an error:', error, errorInfo);
+    this.setState({
+      error: error,
+      errorInfo: errorInfo,
+    });
+  }
+
+  handleReset = () => {
+    this.setState({ hasError: false, error: null, errorInfo: null });
+  };
+
+  render() {
+    if (this.state.hasError) {
+      // Fallback UI
+      return (
+        <div className="error-boundary" role="alert">
+          <div className="error-content">
+            <h2>⚠️ Something went wrong</h2>
+            <p>
+              We encountered an unexpected error. Please try refreshing the page or contact support
+              if the issue persists.
+            </p>
+            <button className="btn primary" onClick={this.handleReset} aria-label="Try again">
+              Try Again
+            </button>
+            <button
+              className="btn secondary"
+              onClick={() => window.location.reload()}
+              aria-label="Refresh page"
+            >
+              Refresh Page
+            </button>
+            {process.env.NODE_ENV === 'development' && (
+              <details className="error-details">
+                <summary>Error Details (Development)</summary>
+                <pre>{this.state.error && this.state.error.toString()}</pre>
+                <pre>{this.state.errorInfo.componentStack}</pre>
+              </details>
+            )}
+          </div>
+        </div>
+      );
+    }
+
+    return this.props.children;
+  }
+}
+
+export default ErrorBoundary;
